ET/ACIA/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.18.0
openai>=1.14.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
requests>=2.31.0
openpyxl>=3.1.2
